+++ README.md
@@ -1,10 +1,8 @@
-# AutoDone-AI
+# AI-Completer
 
 A totally automatic AI to interact with environments.
 
-**At Developping.**
-
-Now it supports OpenAI chat only.
+Now it supports OpenAI API only for model.
 
 ## Setup
 
@@ -21,20 +19,21 @@ cd AutoDone-AI
 
 You can use either vim or your editor to modify the file `config-example.json`, and rename it to `config.json`
 
-3. Run
+3. Run and enjoy yourself
 
 ```shell
-python3 -m autodone
+python3 -m aicompleter
 ```
+
+This normally will start a conversation with ChatGPT, we're adding more supports for this program.
 
 ## To-do List
 
-We are adding more support to this program and 
-- [x] Add Multi-Group Support
+We are adding more support to this program.
 - [ ] Add ChatGPT command support
 - [ ] Add POE support(for plugins)
 - [ ] Add logical structure for automatic task (important)
 - [ ] Add OneBot operation support (optioanl)
-- [x] Add System Interface (partial)
-- [x] Add User Controls and Group Controls
-- [ ] Add Force awaitable for command (and relative return value) <br> for calling more conviniently and catch error
+- [x] Add Force awaitable for command (and relative return value) <br> for calling more conviniently and catch error
+- [ ] Add Commands Intergation with AI model.
+
