+++ aicompleter/session/__init__.py
@@ -1,6 +1,6 @@
 '''
 This package contains the Session class, which is the main class of the
-autodone package. It is used to store the state of the session, including
+aicompleter package. It is used to store the state of the session, including
 the HTTP session, the conversation, the roles, and the extra information.
 '''
 from .base import MultiContent,Content,Text,Image,Audio,Session,Message,MessageQueue
